requirements.txt
absl-py==1.3.0
accelerate==0.14.0
aiohttp==3.8.3
aiosignal==1.3.1
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
contourpy==1.0.6
cycler==0.11.0
datasets==2.6.1
debugpy==1.6.3
decorator==5.1.1
diffusers==0.4.1
dill==0.3.5.1
entrypoints==0.4
executing==1.2.0
expecttest==0.1.3
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.10.0
ftfy==6.1.1
future==0.18.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.10.1
idna==3.3
importlib-metadata==5.0.0
install==1.3.5
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==5.0.0
kiwisolver==1.4.4
llvmlite==0.39.1
lmdb==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
modelcards==0.1.6
multidict==6.0.2
multiprocess==0.70.13
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.20.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.6.0.66
openslide-python==1.2.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-lz4framed==0.14.0
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pynndescent==0.5.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.6.3
six==1.16.0
stack-data==0.6.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch @ file:///dev/shm/pytorch-v1.11.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.24.0
types-dataclasses==0.6.4
typing_extensions==4.1.1
umap-learn==0.5.3
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.2.2
xxhash==3.1.0
yarl==1.8.1
zipp==3.10.0
